rename the folder 'input' to 'inputs', 'RadioInput' to 'Radio' and 'file.tsx' to 'file-input.tsx'.
