--- src/page-modules/contact/components/input/index.ts
@@ -1,1 +1,0 @@
-export { default as Input } from './input';
